requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.1.8
colorama==0.4.6
databases==0.9.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic-core==2.33.1
python-dotenv==1.0.1
ruff==0.11.7
sniffio==1.3.1
sqlalchemy==2.0.31
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
pydantic-settings==2.5.2
apscheduler==3.10.4